+++ src/app/(pages)/Resources/[id]/page.jsx
@@ -3,11 +3,11 @@
 import { useEffect } from 'react';
 import { useRouter, useParams } from 'next/navigation';
 import resources from '../resources';
-import { cppContent } from '@/lib/cpp-content';
-import { dsaContent } from '@/lib/dsa-content';
+// import { cppContent } from '@/lib/cpp-content';
+// import { dsaContent } from '@/lib/dsa-content';
 // import { cpContent } from '@/lib/competitive_programming-content';
 // import { cpTrackContent } from '@/lib/competitive_programming-track-content';
-import { trackContents } from '@/lib/track-contents';
+// import { trackContents } from '@/lib/track-contents';
 import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
 import { Card, CardHeader, CardTitle, CardContent } from "@/components/ui/card";
 import { Button } from "@/components/ui/button";
@@ -108,13 +108,12 @@ const isCompetitiveProgrammingTrack = resource.id === 6;
 // Get appropriate content based on track type
 const getTrackContent = () => {
     if (!resource) return null;
-    return trackContents[resource.id]?.content || null;
+    return resource.content || null;
 };
 
-// Get track-specific info
 const getTrackInfo = () => {
     if (!resource) return null;
-    return trackContents[resource.id]?.trackInfo || null;
+    return resource.trackInfo || null;
 };
 
 const content = getTrackContent();
@@ -143,8 +142,8 @@ const isLearningTrack = content !== null; // Single declaration of isLearningTra
                             <TabsTrigger value="practice">Practice</TabsTrigger>
                         </TabsList>
 
-                        {/* Introduction Tab */}
-                        <TabsContent value="introduction">
+                      {/* Introduction Tab */}
+<TabsContent value="introduction">
     <Card>
         <CardHeader>
             <CardTitle>{resource.title}</CardTitle>
@@ -154,38 +153,40 @@ const isLearningTrack = content !== null; // Single declaration of isLearningTra
                 <p className="text-gray-700 text-lg leading-relaxed">
                     {resource.extendedContent}
                 </p>
-                {trackInfo && (
-                    <>
-                        <div className="mt-6">
-                            <h3 className="text-xl font-semibold mb-3">Prerequisites</h3>
-                            <ul className="list-disc pl-5 space-y-2">
-                                {trackInfo.prerequisites.map((prereq, index) => (
-                                    <li key={index} className="text-gray-600">{prereq}</li>
-                                ))}
-                            </ul>
+                {trackInfo?.prerequisites && (
+                    <div className="mt-6">
+                        <h3 className="text-xl font-semibold mb-3">Prerequisites</h3>
+                        <ul className="list-disc pl-5 space-y-2">
+                            {trackInfo.prerequisites.map((prereq, index) => (
+                                <li key={index} className="text-gray-600">{prereq}</li>
+                            ))}
+                        </ul>
+                    </div>
+                )}
+                {trackInfo?.outcomes && (
+                    <div className="mt-6">
+                        <h3 className="text-xl font-semibold mb-3">Learning Outcomes</h3>
+                        <ul className="list-disc pl-5 space-y-2">
+                            {trackInfo.outcomes.map((outcome, index) => (
+                                <li key={index} className="text-gray-600">{outcome}</li>
+                            ))}
+                        </ul>
+                    </div>
+                )}
+                {trackInfo?.sections && (
+                    <div className="mt-6">
+                        <h3 className="text-xl font-semibold mb-3">Course Sections</h3>
+                        <div className="space-y-4">
+                            {trackInfo.sections.map((section, index) => (
+                                <div key={index} className="bg-gray-50 p-4 rounded-lg">
+                                    <h4 className="font-semibold text-blue-600 mb-2">
+                                        {section.title}
+                                    </h4>
+                                    <p className="text-gray-600">{section.content}</p>
+                                </div>
+                            ))}
                         </div>
-                        <div className="mt-6">
-                            <h3 className="text-xl font-semibold mb-3">Learning Outcomes</h3>
-                            <ul className="list-disc pl-5 space-y-2">
-                                {trackInfo.outcomes.map((outcome, index) => (
-                                    <li key={index} className="text-gray-600">{outcome}</li>
-                                ))}
-                            </ul>
-                        </div>
-                        <div className="mt-6">
-                            <h3 className="text-xl font-semibold mb-3">Course Sections</h3>
-                            <div className="space-y-4">
-                                {trackInfo.sections.map((section, index) => (
-                                    <div key={index} className="bg-gray-50 p-4 rounded-lg">
-                                        <h4 className="font-semibold text-blue-600 mb-2">
-                                            {section.title}
-                                        </h4>
-                                        <p className="text-gray-600">{section.content}</p>
-                                    </div>
-                                ))}
-                            </div>
-                        </div>
-                    </>
+                    </div>
                 )}
             </div>
         </CardContent>
